Moved tests into proper file
